+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased 2.20.0] - 2021-08-07
 ### Added
 - Updated status bar msg if repo is not connected. Clicking on it triggers Init process
+- Applied limit on retry putLogEvent
 
 ## [2.19.0] - 2021-08-07
 ### Added